Refactoring - moved key navigation into git-base-widget
