+++ packages/git/src/browser/history/git-history-widget.ts
@@ -15,14 +15,10 @@ import { GIT_HISTORY, GIT_HISTORY_MAX_COUNT } from './git-history-contribution';
 import { LabelProvider } from '@theia/core/lib/browser/label-provider';
 import { GitRepositoryProvider } from '../git-repository-provider';
 import { GitFileStatus, Git, GitFileChange } from '../../common';
-import { GitBaseWidget } from "../git-base-widget";
-import { GitFileChangeNode } from "../git-widget";
+import { GitBaseWidget, SelectDirection, GitFileChangeNode } from "../git-base-widget";
 import { SelectionService } from "@theia/core";
-import { Message } from "@phosphor/messaging";
-import { ElementExt } from "@phosphor/domutils";
 import { FileSystem } from "@theia/filesystem/lib/common";
 import { EditorWidget } from "@theia/editor/lib/browser";
-import { Key } from "@theia/core/lib/browser/keys";
 
 export interface GitCommitNode {
     readonly authorName: string;
@@ -43,19 +39,15 @@ export namespace GitCommitNode {
     }
 }
 
-export enum SelectDirection {
-    NEXT, PREVIOUS
-}
-
 export type GitHistoryListNode = (GitCommitNode | GitFileChangeNode);
 
 @injectable()
-export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
+export class GitHistoryWidget extends GitBaseWidget<GitHistoryListNode> implements StatefulWidget {
     protected options: Git.Options.Log;
     protected commits: GitCommitNode[];
-    protected historyList: GitHistoryListNode[];
     protected ready: boolean;
     protected singleFileMode: boolean;
+    protected readonly scrollContainer = 'git-history-list-container';
 
     constructor(
         @inject(GitRepositoryProvider) protected readonly repositoryProvider: GitRepositoryProvider,
@@ -64,27 +56,10 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
         @inject(SelectionService) protected readonly selectionService: SelectionService,
         @inject(FileSystem) protected readonly fileSystem: FileSystem,
         @inject(Git) protected readonly git: Git) {
-        super(repositoryProvider, labelProvider);
+        super(repositoryProvider, labelProvider, selectionService);
         this.id = GIT_HISTORY;
         this.title.label = "Git History";
         this.addClass('theia-git');
-
-        this.node.tabIndex = 0;
-
-        selectionService.onSelectionChanged((c: GitHistoryListNode) => {
-            c.selected = true;
-            this.update();
-        });
-    }
-
-    protected onUpdateRequest(msg: Message): void {
-        super.onUpdateRequest(msg);
-
-        const selected = this.node.getElementsByClassName(SELECTED_CLASS)[0];
-        const scrollArea = document.getElementById('git-history-list-container');
-        if (selected && scrollArea) {
-            ElementExt.scrollIntoViewIfNeeded(scrollArea, selected);
-        }
     }
 
     async setContent(options?: Git.Options.Log) {
@@ -166,7 +141,7 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
     }
 
     protected render(): h.Child {
-        this.historyList = [];
+        this.gitNodes = [];
         const containers = [];
         if (this.ready) {
             containers.push(this.renderHistoryHeader());
@@ -204,7 +179,7 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
         const commitList = h.div({ className: "commitList" }, ...theList);
         return h.div({
             className: "listContainer",
-            id: "git-history-list-container",
+            id: this.scrollContainer,
             onscroll: e => {
                 const el = (e.srcElement as HTMLElement);
                 if (el.scrollTop + el.clientHeight > el.scrollHeight - 5) {
@@ -222,7 +197,7 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
     }
 
     protected renderCommit(commit: GitCommitNode): h.Child {
-        this.historyList.push(commit);
+        this.gitNodes.push(commit);
         let expansionToggleIcon = "caret-right";
         if (commit && commit.expanded) {
             expansionToggleIcon = "caret-down";
@@ -274,7 +249,7 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
 
     protected renderFileChangeList(fileChanges: GitFileChangeNode[], commitSha: string): h.Child {
 
-        this.historyList.push(...fileChanges);
+        this.gitNodes.push(...fileChanges);
 
         const files: h.Child[] = [];
 
@@ -315,19 +290,10 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
         return h.div({ className: `gitItem noselect${change.selected ? ' ' + SELECTED_CLASS : ''}` }, ...elements);
     }
 
-    protected onAfterAttach(msg: Message): void {
-        super.onAfterAttach(msg);
-        this.addKeyListener(this.node, Key.ARROW_LEFT, () => this.handleLeft());
-        this.addKeyListener(this.node, Key.ARROW_RIGHT, () => this.handleRight());
-        this.addKeyListener(this.node, Key.ARROW_UP, () => this.handleUp());
-        this.addKeyListener(this.node, Key.ARROW_DOWN, () => this.handleDown());
-        this.addKeyListener(this.node, Key.ENTER, () => this.handleEnter());
-    }
-
     protected handleLeft(): void {
         const selected = this.getSelected();
         if (selected) {
-            if (GitCommitNode.is(selected) && selected.expanded) {
+            if (GitCommitNode.is(selected) && selected.expanded && !this.singleFileMode) {
                 selected.expanded = false;
             } else {
                 this.selectNodeByDirection(SelectDirection.PREVIOUS);
@@ -348,14 +314,6 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
         this.update();
     }
 
-    protected handleUp(): void {
-        this.selectNodeByDirection(SelectDirection.PREVIOUS);
-    }
-
-    protected handleDown(): void {
-        this.selectNodeByDirection(SelectDirection.NEXT);
-    }
-
     protected handleEnter(): void {
         const selected = this.getSelected();
         if (selected) {
@@ -372,34 +330,6 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
         this.update();
     }
 
-    protected onActivateRequest(msg: Message): void {
-        super.onActivateRequest(msg);
-        this.node.focus();
-    }
-
-    protected getSelected(): GitHistoryListNode | undefined {
-        return this.historyList ? this.historyList.find(c => c.selected || false) : undefined;
-    }
-
-    protected selectNode(node: GitHistoryListNode) {
-        const n = this.getSelected();
-        if (n) {
-            n.selected = false;
-        }
-        this.selectionService.selection = node;
-    }
-
-    protected selectNodeByDirection(dir: SelectDirection) {
-        const selIdx = this.historyList.findIndex(c => c.selected || false);
-        let nodeIdx = selIdx;
-        if (dir === SelectDirection.NEXT && selIdx < this.historyList.length - 1) {
-            nodeIdx = selIdx + 1;
-        } else if (dir === SelectDirection.PREVIOUS && selIdx > 0) {
-            nodeIdx = selIdx - 1;
-        }
-        this.selectNode(this.historyList[nodeIdx]);
-    }
-
     protected openFile(change: GitFileChange, commitSha: string) {
         const uri: URI = new URI(change.uri);
         let fromURI = change.oldUri ? new URI(change.oldUri) : uri; // set oldUri on renamed and copied
@@ -413,9 +343,7 @@ export class GitHistoryWidget extends GitBaseWidget implements StatefulWidget {
         } else {
             uriToOpen = DiffUris.encode(fromURI, toURI, uri.displayName);
         }
-        open(this.openerService, uriToOpen).then((e: EditorWidget) => {
-
-        });
+        open(this.openerService, uriToOpen, { activate: false });
     }
 
 }
